Orderservice/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy<2.0
flask-migrate==3.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
flask-cors==3.0.10
flask-jwt-extended==4.3.1
requests==2.28.1
email-validator==1.1.3